+++ const/BomSpells.lua
@@ -21,15 +21,18 @@ local BOM_PHYSICAL_CLASSES = { "HUNTER", "ROGUE", "SHAMAN", "WARRIOR", "DRUID" }
 ---@field classes table - list of target classes which are shown as toggle boxes to enable cast per class
 ---@field default boolean - whether the spell auto-cast is enabled by default
 ---@field groupDuration number - buff duration for group buff in seconds
----@field groupFamily table - family of group buffs which are mutually exclusive
+---@field groupFamily table - family of group buff spell ids which are mutually exclusive
 ---@field groupId number - spell id for group buff
+---@field hasCD boolean - there's a cooldown on this spell
 ---@field isAura boolean - true if the buff affects others in radius, and not a target buff
 ---@field isTracking boolean - true if the buff is tracking of some resource or enemy
 ---@field isWeapon boolean - the buff is a temporary weapon enchant on user's weapons
 ---@field item number - buff is granted by an item in user's bag
+---@field itemLock table - item ids
+---@field NeededGroupItem table - reagent item ids for group buff
 ---@field onlyUsableFor table - list of classes which only can see this buff (hidden for others)
 ---@field singleDuration number - buff duration for single buff in seconds
----@field singleFamily table - family of single buffs which are mutually exclusive
+---@field singleFamily table - family of single buff spell ids which are mutually exclusive
 ---@field singleId number - spell id for single buff
 BOM.AllBuffomatSpells = {
   ----PRIEST----
@@ -285,7 +288,7 @@ BOM.AllBuffomatSpells = {
     onlyUsableFor = BOM_MANA_CLASSES }, --Monster Omelet x
   { singleId = 25661, item = 21023, isBuff = true, default = false }, --Dirge's Kickin' Chimaerok Chops x
   { singleId      = 18141, item = 13813, isBuff = true, default = false,
-    onlyUsableFor = BOM_MANA_CLASSES }, --Blessed Sunfruit Juice +Strength
+    onlyUsableFor = BOM_MELEE_CLASSES }, --Blessed Sunfruit Juice +Strength
   { singleId      = 18125, item = 13810, isBuff = true, default = false,
     onlyUsableFor = BOM_MANA_CLASSES }, --Blessed Sunfruit +Spirit
   { singleId      = 22790, item = 18284, isBuff = true, default = false,